+++ CHANGELOG.md
@@ -1,16 +1,32 @@
 # Version History
 
-- ????-??-?? x.x.x
+- 2025-??-?? 4.1.24
+    - MatchingLibrary
+        - deprecated spectra lacking provenance
+        - added fresh 785X
     - docs
         - Update install instructions for RPi 4 and PySide6 
-
-- 2024-02-11 4.1.22
+- 2025-02-20 4.1.23
+    - plugins
+        - Analysis.Stats 
+            - allow user to dynamically change multi_devices
+        - Analysis.Peakfinding 
+            - add params 
+            - add copy button
+        - Analysis.Tail
+            - simplified
+            - add DidYouKnow
+    - Marquee 
+        - made threadsafe for plugins
+        - removed toast
+- 2025-02-11 4.1.22
+>>>>>>> main
     - updated Wasatch.PY to 2.2.17 to resolve broken Batch Collection and TakeOne buttons
     - plugins
         - RnD.ColorMatching: more robust handling of missing calibration
         - support error Marquee messages
     - Marquee: re-use single QTimer, to reduce start/stop events
-- 2024-02-06 4.1.21
+- 2025-02-06 4.1.21
     - [#511] support graphing of loaded spectra with only wavenumber axis
     - added counter to Analysis/Stats
     - initial Analysis/Tail draft